Add global video playback block for a slow device you can select based on UserAgent.
As test I chose the add Amazon Kindle in there. It does never trigger on it as the browser is way to minimalistic to even trifgger my script on these devices, but that way you can hand edit it to your likes.

Wanna change it to some other kind of device? There are two ways to let it pass on the User Agent on the too slow device. As Debug Console output or as lil Altert Message box. Choose one and find something in the Agent you can select as detection method.